requirements.txt
# discord python package, and dotenv
discord.py
python-dotenv
aiohttp
logging
os
asyncio
datetime
random
json